Began adding shop items.
The 'shop' is going to be a bit different in this iteration. FTB Money
no longer has the option to gate trades behind quest completion, and FTB
Quests no longer has the option to make repeatable quests, so neither is
suitable for the shop.

Instead, items can be 'bought' by crafting coins with an existing item
to duplicate it. It's a bit of a combination of Equivalent Exchange and
UU-Matter.